Feat #17430: Swing percentage feature added
change swingRatio type to float and update related UI components,
and updated test for the feature.

+++ src/engraving/types/types.h
@@ -1142,7 +1142,7 @@ static inline Key operator-=(Key& a, const Key& b) { return a = Key(static_cast<
 
 struct SwingParameters {
     int swingUnit = 0;
-    int swingRatio = 0;
+    float swingRatio = 0;
 
     bool isOn() const { return swingUnit != 0; }
 };
